Back actions refactoring
